+++ resources/views/filament/pages/general-settings.blade.php
@@ -2,6 +2,10 @@
     <x-filament-panels::form wire:submit="save">
         {{ $this->form }}
 
-        <x-filament-panels::form.actions :actions="$this->getCachedFormActions()" />
+        <div class="flex justify-start">
+            <x-filament::button type="submit">
+                Save
+            </x-filament::button>
+        </div>
     </x-filament-panels::form>
 </x-filament-panels::page>